Remove async-await and promisify module save
Async-await only works in the latest version of node. Even my computer
has a version that it doesn't work. I think for the sake of working
in a wider audience, we should avoid them.

I also separated the Save module into smaller functions to make it easier
to spot bugs and understand the code.